Replace Test Functional->Integration for Provider::AdminControllerTest
